Fix Precompile Address Width

+++ modules/passkey/test/4337/WebAuthnSigner.spec.ts
@@ -110,7 +110,7 @@ describe('WebAuthn Signers [@4337]', () => {
           { type: 'address', name: 'signerFactory' },
           { type: 'uint256', name: 'signerX' },
           { type: 'uint256', name: 'signerY' },
-          { type: 'uint192', name: 'signerVerifiers' },
+          { type: 'uint176', name: 'signerVerifiers' },
           { type: 'address', name: 'setupTo' },
           { type: 'bytes', name: 'setupData' },
           { type: 'address', name: 'fallbackHandler' },
